+++ artifact_backend/artifactapp/views/auth.py
@@ -1,0 +1,43 @@
+from django.contrib.auth import authenticate, login, logout
+from rest_framework import status, views
+from rest_framework.response import Response
+from ..serializers import UserSerializer, LoginSerializer
+
+
+class SignupView(views.APIView):
+    def post(self, request):
+        serializer = UserSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+class LoginView(views.APIView):
+    def post(self, request):
+        serializer = LoginSerializer(data=request.data)
+        if serializer.is_valid():
+            user = authenticate(username=serializer.validated_data['username'],
+                                password=serializer.validated_data['password'])
+            if user:
+                login(request, user)
+                return Response({"message": "Login successful"}, status=status.HTTP_200_OK)
+            return Response({"error": "Invalid credentials"}, status=status.HTTP_401_UNAUTHORIZED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+class LogoutView(views.APIView):
+    def post(self, request):
+        logout(request)
+        return Response({"message": "Logged out successfully"}, status=status.HTTP_200_OK)
+
+
+class HealthView(views.APIView):
+    def get(self, request):
+        return Response(
+            {
+                "status": 200,
+                "message": "The app is up and running."
+            },
+            status=status.HTTP_200_OK
+        )
